Create utl_Find_Missing_Objects_Between_Schemas.sql
SQL Server 2016 and above T-SQL to find the missing objects between the specified old and new schemas.  Presently, the logic is hardwired to handle just two each of the old and new schemas. Obviously, a future enhancement would be to handle more than two each of the old and new schemas.